fix path issues related to folder reorganisation
