+++ policy-engine-modules/policy-engine-ui/frontend/src/components/policy-engine-app/attributes/Attributes.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect, /*useRef*/ } from 'react';
-import "../static/css/Zonesoft.css"
+import "../../../static/css/Zonesoft.css"
 import * as DataService from "./DataService";
 
 function Attributes(props) {
